requirements.txt
functions-framework
flask
google-cloud-error-reporting
MarkupSafe
google-cloud-bigquery
